+++ CHANGELOG.md
@@ -5,8 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-## Changed
+### Changed
 - Update pyproject.toml with readme / repository info
+
+### Removed
+- `structlog` dependency (all logging using vanilla `logging` now)
 
 ## [0.1.4] - 2022-08-03
 ### Added